unified resize handler where possible to increase code reuse
Changelog: None
